Use ShopperContextInterface
The `setting` TWIG function will always return null when the `sylius` global TWIG variable is not an instance of `\Sylius\Component\Core\Context\ShopperContext`, i.e. when a custom context is used. Instead, check if it implements `Sylius\Component\Core\Context\ShopperContextInterface`.
Always try to rely on abstractions in libraries.